* Changes to status codes     - removed the VAL_AC_TRUSTED_ZONE and VAL_TRUSTED_ZONE codes     - renamed *PROVABLY_INSECURE* to *PINSECURE*     - renamed VAL_LOCAL_ANSWER to VAL_OOB_ANSWER * Changes to types     - using u_char in place of u_int8_t to represent character buffers     - using size_t where appropriate     - Decoupled val_rrset_rec from other internal structures. Moved members     from older val_rrset_digested structure to rrset_rec structure     - No longer store complete header information in val_rrset_rec; only keep     rcode information * Changes to function prototypes     - using higher order types instead of u_int* variants for exported     functions     - val_add_valpolicy() now takes a void * as the second argument     - No longer pass OTW format domain names for val_resolve_and_check() * policy-related     - No longer define a "trusted" zone security expectation since there is no     way to give it special treatment in the high-level API; the "ignore"     condition subsumes the "trusted" condition.     - Define new libval_policy_definition_t structure to encapsulate the     libval-specific policy definition parameters in val_add_valpolicy()
git-svn-id: https://www.dnssec-tools.org/svn/dnssec-tools/trunk@4305 c4029f9c-b713-0410-a448-fae9e31a314a

+++ dnssec-tools/validator/libsres/res_debug.c
@@ -257,7 +257,7 @@ do_section(ns_msg * handle, ns_sect section, int pflag, FILE * file)
  * This is intended to be primarily a debugging routine.
  */
 void
-libsres_pquery(const u_char * msg, int len, FILE * file)
+libsres_pquery(const u_char * msg, size_t len, FILE * file)
 {
     ns_msg          handle;
     int             qdcount, ancount, nscount, arcount;
